bark : move cpp structs to bark.cpp and make C header
