+++ bark.h
@@ -1,276 +1,208 @@
-#include <map>
-#include <random>
-#include <thread>
-#include <vector>
-
 #include "encodec.h"
 #include "ggml-backend.h"
 #include "ggml.h"
 
-enum class bark_verbosity_level {
-    LOW = 0,
-    MEDIUM = 1,
-    HIGH = 2,
-};
+#ifdef __cplusplus
+extern "C" {
+#endif
+    enum class bark_verbosity_level {
+        LOW = 0,
+        MEDIUM = 1,
+        HIGH = 2,
+    };
 
-typedef int32_t bark_token;
+    typedef int32_t bark_token;
 
-typedef std::vector<int32_t> bark_sequence;
-typedef std::vector<std::vector<int32_t>> bark_codes;
+    typedef std::vector<int32_t> bark_sequence;
+    typedef std::vector<std::vector<int32_t>> bark_codes;
 
-struct gpt_hparams {
-    int32_t n_in_vocab;
-    int32_t n_out_vocab;
-    int32_t n_layer;
-    int32_t n_head;
-    int32_t n_embd;
-    int32_t block_size;
-    int32_t n_lm_heads;
-    int32_t n_wtes;
-    int32_t ftype;
-    int32_t bias;
+    struct gpt_hparams {
+        int32_t n_in_vocab;
+        int32_t n_out_vocab;
+        int32_t n_layer;
+        int32_t n_head;
+        int32_t n_embd;
+        int32_t block_size;
+        int32_t n_lm_heads;
+        int32_t n_wtes;
+        int32_t ftype;
+        int32_t bias;
 
-    int32_t n_codes_given = 1;
-};
+        int32_t n_codes_given = 1;
+    };
 
-struct bark_vocab {
-    using id = int32_t;
-    using token = std::string;
+    struct bark_context;
+    struct bark_model;
 
-    std::map<token, id> token_to_id;
-    std::map<id, token> id_to_token;
-};
+    // Holds the vocabulary for the semantic encoder
+    struct bark_vocab;
 
-struct gpt_layer {
-    // normalization
-    struct ggml_tensor *ln_1_g;
-    struct ggml_tensor *ln_1_b;
+    // Define the GPT architecture for the 3 encoders
+    struct gpt_model;
 
-    struct ggml_tensor *ln_2_g;
-    struct ggml_tensor *ln_2_b;
+    struct bark_statistics {
+        // Time to load model weights
+        int64_t t_load_us = 0;
+        // Time to generate audio
+        int64_t t_eval_us = 0;
 
-    // attention
-    struct ggml_tensor *c_attn_attn_w;
-    struct ggml_tensor *c_attn_attn_b;
+        int64_t t_semantic_us = 0;
+        int64_t t_coarse_us   = 0;
+        int64_t t_fine_us     = 0;
 
-    struct ggml_tensor *c_attn_proj_w;
-    struct ggml_tensor *c_attn_proj_b;
+        int32_t n_sample_semantic = 0;
+        int32_t n_sample_coarse   = 0;
+        int32_t n_sample_fine     = 0;
+    };
 
-    // mlp
-    struct ggml_tensor *c_mlp_fc_w;
-    struct ggml_tensor *c_mlp_fc_b;
+    struct bark_context_params {
+        // Verbosity level
+        bark_verbosity_level verbosity;
 
-    struct ggml_tensor *c_mlp_proj_w;
-    struct ggml_tensor *c_mlp_proj_b;
-};
+        // Temperature for sampling (text and coarse encoders)
+        float temp;
+        // Temperature for sampling (fine encoder)
+        float fine_temp;
 
-struct gpt_model {
-    gpt_hparams hparams;
+        // Minimum probability for EOS token (text encoder)
+        float min_eos_p;
+        // Sliding window size for coarse encoder
+        int32_t sliding_window_size;
+        // Max history for coarse encoder
+        int32_t max_coarse_history;
 
-    // normalization
-    struct ggml_tensor *ln_f_g;
-    struct ggml_tensor *ln_f_b;
+        // Sample rate
+        int32_t sample_rate;
+        // Target bandwidth
+        int32_t target_bandwidth;
 
-    struct ggml_tensor *wpe;                     //  position embedding
-    std::vector<struct ggml_tensor *> wtes;      //     token embedding
-    std::vector<struct ggml_tensor *> lm_heads;  // language model head
+        // CLS token ID
+        int32_t cls_token_id;
+        // SEP token ID
+        int32_t sep_token_id;
 
-    std::vector<gpt_layer> layers;
+        // Maximum number of semantic tokens to generate
+        int32_t n_steps_text_encoder;
 
-    // key + value memory
-    struct ggml_tensor *memory_k;
-    struct ggml_tensor *memory_v;
+        // Text PAD token ID
+        int32_t text_pad_token;
+        // Text encoding offset
+        int32_t text_encoding_offset;
 
-    struct ggml_context *ctx;
+        // Semantic frequency rate
+        float semantic_rate_hz;
+        // Semantic PAD token ID
+        int32_t semantic_pad_token;
+        // Vocabulary size in semantic encoder
+        int32_t semantic_vocab_size;
+        // Semantic infernce token ID
+        int32_t semantic_infer_token;
 
-    ggml_backend_t backend = NULL;
+        // Coarse frequency rate
+        float coarse_rate_hz;
+        // Coarse infer token ID
+        int32_t coarse_infer_token;
+        // Coarse semantic pad token ID
+        int32_t coarse_semantic_pad_token;
 
-    ggml_backend_buffer_t buffer_w;
-    ggml_backend_buffer_t buffer_kv;
+        // Number of codebooks in coarse encoder
+        int32_t n_coarse_codebooks;
+        // Number of codebooks in fine encoder
+        int32_t n_fine_codebooks;
+        // Dimension of the codes
+        int32_t codebook_size;
+    };
 
-    std::map<std::string, struct ggml_tensor *> tensors;
+    /**
+     * @brief Returns the default parameters for a bark context.
+     *
+     * @return bark_context_params The default parameters for a bark context.
+     */
+    struct bark_context_params bark_context_default_params(void);
 
-    //
-    int64_t t_sample_us = 0;
-    int64_t t_predict_us = 0;
-    int64_t t_main_us = 0;
+    /**
+     * Loads a BARK model from the specified file path with the given parameters.
+     *
+     * @param model_path The directory path of the bark model to load.
+     * @param verbosity  The verbosity level when loading the model.
+     * @param seed       The seed to use for random number generation.
+     * @return A pointer to the loaded bark model context.
+     */
+    struct bark_context *bark_load_model(
+        const char *model_path,
+        bark_verbosity_level verbosity,
+        uint32_t seed);
 
-    //
-    int64_t n_sample = 0;
+    /**
+     * Generates an audio file from the given text using the specified Bark context.
+     *
+     * @param bctx The Bark context to use for generating the audio.
+     * @param text The text to generate audio from.
+     * @param n_threads The number of threads to use for generating the audio.
+     * @return An integer indicating the success of the audio generation process.
+     */
+    bool bark_generate_audio(
+        struct bark_context *bctx,
+        const char *text,
+        int n_threads);
 
-    //
-    int64_t memsize = 0;
-};
+    /**
+     * Retrieves the audio data generated by the Bark context.
+     *
+     * @param bctx The Bark context to use for generating the audio.
+     * @return A pointer to the audio data generated by the Bark context.
+     */
+    float * bark_get_audio_data(
+        struct bark_context *bctx);
 
-struct bark_model {
-    // encoder
-    gpt_model coarse_model;
-    gpt_model fine_model;
-    gpt_model semantic_model;
+    /**
+     * Retrieves the audio data generated by the Bark context.
+     *
+     * @param bctx The Bark context to use for generating the audio.
+     * @return The size of the audio data generated by the Bark context.
+     */
+    int bark_get_audio_data_size(
+        struct bark_context *bctx);
 
-    // vocab
-    bark_vocab vocab;
-};
+    /**
+     * Retrieves the statistics of the last audio generation round.
+     *
+     * @param bctx The Bark context to use for generating the audio.
+     * @return A struct containing the statistics of the last audio generation round.
+     */
+    const bark_statistics * bark_get_statistics(
+        struct bark_context *bctx);
 
-struct bark_context_params {
-    // Verbosity level
-    bark_verbosity_level verbosity;
+    /**
+     * Reset the statistics of the last audio generation round.
+     *
+     * @param bctx The Bark context to use for generating the audio.
+     * @return A struct containing the statistics of the last audio generation round.
+     */
+    void bark_reset_statistics(
+        struct bark_context *bctx);
 
-    // Temperature for sampling (text and coarse encoders)
-    float temp;
-    // Temperature for sampling (fine encoder)
-    float fine_temp;
+    /**
+     * Quantizes a bark model and saves the result to a file.
+     *
+     * @param fname_inp The name of the input file containing the BARK model.
+     * @param fname_out The name of the output file to save the quantized model to.
+     * @param ftype The type of the model's floating-point values.
+     * @return True if the model was successfully quantized and saved, false otherwise.
+     */
+    bool bark_model_quantize(
+        const char * fname_inp,
+        const char * fname_out,
+        ggml_ftype ftype);
 
-    // Minimum probability for EOS token (text encoder)
-    float min_eos_p;
-    // Sliding window size for coarse encoder
-    int32_t sliding_window_size;
-    // Max history for coarse encoder
-    int32_t max_coarse_history;
+    /**
+     * @brief Frees the memory allocated for a bark context.
+     *
+     * @param bctx The bark context to free.
+     */
+    void bark_free(
+        struct bark_context *bctx);
 
-    // Sample rate
-    int32_t sample_rate;
-    // Target bandwidth
-    int32_t target_bandwidth;
-
-    // CLS token ID
-    int32_t cls_token_id;
-    // SEP token ID
-    int32_t sep_token_id;
-
-    // Maximum number of semantic tokens to generate
-    int32_t n_steps_text_encoder;
-
-    // Text PAD token ID
-    int32_t text_pad_token;
-    // Text encoding offset
-    int32_t text_encoding_offset;
-
-    // Semantic frequency rate
-    float semantic_rate_hz;
-    // Semantic PAD token ID
-    int32_t semantic_pad_token;
-    // Vocabulary size in semantic encoder
-    int32_t semantic_vocab_size;
-    // Semantic infernce token ID
-    int32_t semantic_infer_token;
-
-    // Coarse frequency rate
-    float coarse_rate_hz;
-    // Coarse infer token ID
-    int32_t coarse_infer_token;
-    // Coarse semantic pad token ID
-    int32_t coarse_semantic_pad_token;
-
-    // Number of codebooks in coarse encoder
-    int32_t n_coarse_codebooks;
-    // Number of codebooks in fine encoder
-    int32_t n_fine_codebooks;
-    // Dimension of the codes
-    int32_t codebook_size;
-};
-
-struct bark_context {
-    bark_model text_model;
-
-    struct encodec_context *encodec_ctx;
-
-    // buffer for model evaluation
-    ggml_backend_buffer_t buf_compute;
-
-    // custom allocator
-    struct ggml_allocr *allocr = NULL;
-    int n_gpu_layers = 0;
-
-    std::mt19937 rng;
-
-    bark_sequence tokens;
-    bark_sequence semantic_tokens;
-
-    bark_codes coarse_tokens;
-    bark_codes fine_tokens;
-
-    std::vector<float> audio_arr;
-
-    // hyperparameters
-    bark_context_params params;
-
-    // statistics
-    int64_t t_load_us = 0;
-    int64_t t_eval_us = 0;
-
-    // encodec parameters
-    std::string encodec_model_path;
-};
-
-/**
- * @brief Returns the default parameters for a bark context.
- *
- * @return bark_context_params The default parameters for a bark context.
- */
-struct bark_context_params bark_context_default_params(void);
-
-/**
- * Loads a BARK model from the specified file path with the given parameters.
- *
- * @param model_path The directory path of the bark model to load.
- * @param verbosity  The verbosity level when loading the model.
- * @param seed       The seed to use for random number generation.
- * @return A pointer to the loaded bark model context.
- */
-struct bark_context *bark_load_model(
-    const char *model_path,
-    bark_verbosity_level verbosity,
-    uint32_t seed);
-
-/**
- * Generates an audio file from the given text using the specified Bark context.
- *
- * @param bctx The Bark context to use for generating the audio.
- * @param text The text to generate audio from.
- * @param n_threads The number of threads to use for generating the audio.
- * @return An integer indicating the success of the audio generation process.
- */
-bool bark_generate_audio(
-    struct bark_context *bctx,
-    const char *text,
-    int n_threads);
-
-/**
- * Retrieves the audio data generated by the Bark context.
- *
- * @param bctx The Bark context to use for generating the audio.
- * @return A pointer to the audio data generated by the Bark context.
- */
-float * bark_get_audio_data(
-    struct bark_context *bctx);
-
-/**
- * Retrieves the audio data generated by the Bark context.
- *
- * @param bctx The Bark context to use for generating the audio.
- * @return The size of the audio data generated by the Bark context.
- */
-int bark_get_audio_data_size(
-    struct bark_context *bctx);
-
-/**
- * Quantizes a bark model and saves the result to a file.
- *
- * @param fname_inp The name of the input file containing the BARK model.
- * @param fname_out The name of the output file to save the quantized model to.
- * @param ftype The type of the model's floating-point values.
- * @return True if the model was successfully quantized and saved, false otherwise.
- */
-bool bark_model_quantize(
-    const std::string &fname_inp,
-    const std::string &fname_out,
-    ggml_ftype ftype);
-
-/**
- * @brief Frees the memory allocated for a bark context.
- *
- * @param bctx The bark context to free.
- */
-void bark_free(
-    struct bark_context *bctx);
+#ifdef __cplusplus
+}
+#endif
